Update ShiftsController.cs
Hotfix for changing Long to Int in assignAvailability. Was causing 500...

+++ Planner/Controllers/ShiftsController.cs
@@ -183,7 +183,7 @@ namespace Planner.Controllers
 
         // POST: Shifts/assignShift
         [HttpPost]
-        public async Task<IActionResult> assignAvailability(int shiftId, long availabilityId)
+        public async Task<IActionResult> assignAvailability(int shiftId, int availabilityId)
         {
             try
             {
